DockerTest/dockerApp/serializers.py
from rest_framework import serializers
from .models import Account, Contact

class AccountSerializer(serializers.ModelSerializer):
    class Meta:
        model = Account
        fields = "__all__"

class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = "__all__"